part4-complete

+++ part4/bloglist/controllers/blogs.js
@@ -1,8 +1,10 @@
 const blogsRouter = require("express").Router();
+const jwt = require("jsonwebtoken");
 const Blog = require("../models/blog");
+const User = require("../models/user");
 
 blogsRouter.get("/", async (req, res) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
   res.json(blogs);
 });
 
@@ -17,30 +19,36 @@ blogsRouter.get("/:id", async (req, res) => {
 
 blogsRouter.post("/", async (req, res) => {
   const body = req.body;
-  let blog;
+  const user = req.user;
 
-  if (body.hasOwnProperty("likes")) {
-    blog = new Blog({
-      title: body.title,
-      author: body.author,
-      url: body.url,
-      likes: body.likes,
-    });
-  } else {
-    blog = new Blog({
-      title: body.title,
-      author: body.author,
-      url: body.url,
-      likes: 0,
-    });
+  if(body.url === undefined || body.title === undefined){
+    return res.status(400).json({error: 'Url or title missing!'})
   }
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes === undefined ? 0 : body.likes,
+    user: user._id,
+  });
 
   const savedBlog = await blog.save();
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
   res.json(savedBlog);
 });
 
 blogsRouter.delete("/:id", async (req, res) => {
-  await Blog.findByIdAndRemove(req.params.id);
+  const user = req.user;
+  const blog = await Blog.findById(req.params.id);
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(req.params.id);
+  } else {
+    return res
+      .status(401)
+      .json({ error: "this blog was not posted by this user" });
+  }
   res.status(204).end();
 });
 
@@ -54,14 +62,10 @@ blogsRouter.put("/:id", async (req, res) => {
     likes: body.likes,
   };
 
-  const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, blog, {new: true});
+  const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, blog, {
+    new: true,
+  });
   res.json(updatedBlog);
-
-  /* Blog.findByIdAndUpdate(req.params.id, blog, { new: true })
-    .then((updatedBlog) => {
-      res.json(updatedBlog);
-    })
-    .catch((error) => next(error)); */
 });
 
 module.exports = blogsRouter;
